Fix broken song downloading from song info page.
Update the slimserver logotype to say slimserver.

Work on the installer script a bit.  Not done.

Added a button that jumps you to the synchronize setting per a user request.

Reset the brightness correctly when we do a firmware update.

Latest firmware fixes a bug where we could have a bogus MAC address if you used
a hardware downloader.

Document IE's annoying click.


git-svn-id: file:///home/awy/mirror/slim/trunk/platforms@286 62299810-d8cb-41fd-93b7-d32162e5a4a4