Michael Herger submitted some patches to the Handheld skin a while back.  Finally got to them.
A few folks have requested the display of the name of the playlist currently playing.  We now display this on the EN, Default and Handheld skins.  If it gets changed,
it adds "(Modified)", if the playlist was created from scratch, it's not displayed at all.

Dave Nanian submitted a patch to create the logs folder if necessary on OSX.


git-svn-id: file:///home/awy/mirror/slim/trunk/platforms@1121 62299810-d8cb-41fd-93b7-d32162e5a4a4