the Artwork cache folder can contain several thousands of files. The Windows installer would check permission on all of these files during the "creating directories" stage of the installation. This lead several users to think the installer had crashed.
Now that we no longer use the Artwork cache folder we should get rid of it to speed up updates.

Please note that you'll not notice any improvement the first time you install this update, as removing the folder will take the same amount of time as checking permissions. But subsequent updates will no longer see this delay.

Bug: 9698
git-svn-id: file:///home/awy/mirror/slim/7.5/branches/embedded/platforms@30396 62299810-d8cb-41fd-93b7-d32162e5a4a4