Debian: Add additional package dependencies
1. Sox

The LMS bundled Sox binary (v14.3.1) has been built with OpenMP support,
and has a dependency on libgomp1. Sox will fail if this library is not
present. This failure is difficult for a user to diagnose.

Sox 14.3.1 was available in Debian Lenny, and specified libgomp1
version >= 4.2.1. So repeating that here.


2. IO:Socket::SSL

LMS does not include a bundled IO:Socket::SSL. Although LMS operates
without the presence of IO:Socket::SSL package, HTTPS streams, etc.
will not be available.

Source package is libio-socket-ssl-perl.


3. Others

libstdc++6
libgcc1

A number of the bundled binaries and CPAN packages depend on these two
libraries.

zlib1g

Image::Scale and Audio::Scan require the zlib library.