Debian: Add package dependency - ca-certificates
A dependency on libio-socket-ssl-perl (IO::Socket::SSL) was added in
February 2018 to ensure that the LMS installation is capable of using
HTTPS.

However, this is not enough to provide seamless operation. At present,
many HTTPS operations will simply fail in the absence of a valid set of
root certificates. In particular, users may be unable to set up their
user account on first time installation, and the extensions downloader
is unlikely to work. 

This change adds Debian’s ‘ca-certificates’ package to the list of
dependencies, to ensure that root certificates are present. HTTPS
operations will then proceed as expected.


Most users will not be impacted by this change, as they will already
have the package installed through other means. But if LMS is set up in
a ‘thin’ Debian installation (e.g. a newly created container), then the
package may not be present.