Debian: Remove package dependency - libgomp1
A dependency on libgomp1 (OpenMP) was added in February 2018 because the
LMS bundled Sox binary (v14.3.1) had been built with OpenMP support,
having a dependency on libgomp1. Sox would fail if this library is not
present.

But the Sox dependency on libgomp1 has been removed following this pull
request in August 2018:

‘Update sox and flac to latest source’
https://github.com/Logitech/slimserver-vendor/pull/58

In particular within this commit:
https://github.com/Logitech/slimserver-vendor/commit/19b35e9f54898befd06c9cc7cbfa44f3491b8f20
wherein is added the configure flag --disable-openmp.


So removing this dependency, as it is no longer required.