Make LMS' httpport configurable through environment variable
The port used in the container needs to be the same as mapped outside, as its part of the slimproto announcement used by players to connect to the server. Therefore a simple mapping from default 9000 internally to 9002 outside would not allow a player to connect to a container run on eg. Synology NAS where port 9000 is occupied.