Debian: Add systemd service unit
This change adds a systemd service unit file to the Debian installation
package.

On a systemd enabled system, this service file will be used in place of
the SysV init scripts that have traditionally used to start LMS. It
offers better logging at LMS start up by eliminating use of the
'squeezeboxserver_safe' script. Other behavioural differences are noted
by way of comments in the unit file.

On a "non-systemd" system, the existing SysV init scripts will be used.

The method used is that published in the Debian systemd packaging wiki:
https://wiki.debian.org/Teams/pkg-systemd/Packaging

The present packaging uses debhelper compatibility level 9, and that
remains unchanged. The required amendments are described in the wiki.

debian/control:
+ Added Build-Depends on dh-systemd (>= 1.5)
+ Also added '${misc:Depends}' and '${misc:Pre-Depends}'. This is not
  referenced in the wiki, but the "systemd helper" may populate these
  fields.

debian/rules:
+ 'dh' is now invoked '--with systemd'

debian/logitechmediaserver.service
+ This is the systemd unit file.