Debian: Add systemd service unit - support older systems without 'init-system-helpers'
The debhelper systemd tools may automatically add a dependency on
package 'init-system-helpers' to the LMS package. (debhelper <= 11 will
do this).

This dependency cannot/may not be met on older Debian based systems
(Debian version <= 7), and in consequence the package would fail to
install on these systems. These systems will not be using 'systemd'.

There continues to be a significant element of the user base that runs
these older systems, and they need to be accommodated.

This change does that by simply removing the dependency from the
generated debian control file, should it be there.

The LMS package will be successfully installed and started regardless of
whether the 'init-system-helpers' package is available. The user of a
'deficient' system will see the following message, or similar, during
installation:

/var/lib/dpkg/info/logitechmediaserver.postinst: 61:
   /var/lib/dpkg/info/logitechmediaserver.postinst: deb-systemd-helper: not found

Nevertheless, LMS will have been installed, and the traditional SysV
init script will be in place.

Package 'init-system-helpers' has been an 'essential' package since
Debian version 9, so modern systems will not be impacted by removing the
dependency.