Debian: systemd service unit - minor modifications
This change modifies the systemd service unit as follows:

+ The systemd stanza 'Group=nogroup' is removed.

What is wanted is that LMS be started with the system default, whatever
that may be. At present the system default is 'nogroup', so the stanza
is a no-op. The stanza should be removed to ensure that the system
sets the (default) group that LMS is launched under.

This will replicate the behaviour of the SysV init script.

+ Language clarification in the comments

By "run time variables" is meant "start up parameters", i.e. those
defined in '/etc/default/logitechmediaserver', and which influence
LMS' start up behaviour.