Weak dependency on perl IO::Socket:SSL in RPM
Added a weak dependency on the perl module IO::Socket::SSL in the RPMK spec file.

Zypper and dnf will automatically install the package providing this module if it is available in the repositories. If not just the logitechmediaserver package will be installed. Yum (pre Centos/RHEL v8) and rpm do not evaluate weak dependencies and thus this addition will not affect installations made with those tools.