Log early startup messages to syslog
This change is designed to replicate the logging behaviour under systemd systems for non-systemd systems (i.e. those using the init.d script). This will ensure that early error messages, before LMS logging is up and running, are available in syslog. Once LMS is up, the default --quiet parameter will ensure that normal logging will not be duplicated in syslog.