Improved config migration
Improved the configuration migration from squeezeboxserver to lyrionmusicserver.
Some plugins (e.g. cdplayer) requires that the user running the music server is in specific groups. The RPM post-script now checks if the user squeezeboxserver is in any other groups than squeezeboxserver, if it is, then the script adds the user lyrionmusicserver to those groups too.

This should give a smother upgrade from LMS 8.5.X to LMS 9.0.0 for those users that use these apps.

Also improved some informational messages and the content of the rebranding README file.